+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+- Added rules precedence check
+
 ## v0.1.0 [2016-04-05]
 
 - First versioned release